generate testing data
This program generates 1000 sorted strings with max len 100000.

For each string, I randomly pick a character from a-z as the first
character of the string.

Then use rand()%MAX_LEN to determine the length of this string.

Fill the string with the char chosen earlier. This char will
increaces its ascii code value with the probability 10/len.
Once the char is about to increase its ascii code, it increaces
by range from 1~4 with the change 19/20, and by range from 1~25
with probability 1/20.

Most of the probability values come from trial and error.
I havn't figure out an efficient way to decide the values.